Fixes some warnings. Removed a couple of calls to nonexistent functions and put in a couple of casts.
git-svn-id: https://frontierkernel.svn.sourceforge.net/svnroot/frontierkernel/Frontier/trunk@1187 4e289971-d10f-0410-980f-25908192c0f7

+++ Common/source/icon.c
@@ -684,7 +684,7 @@ drawiconsequence (Rect r, short firsticon, short lasticon, bigstring bs) {
 		
 		rarrow.left++; /*it appears to need it*/
 				
-		DrawThemePopupArrow (&rarrow, kThemeArrowRight, kThemeArrow9pt, kThemeStateActive, NULL, NULL);
+		DrawThemePopupArrow (&rarrow, kThemeArrowRight, kThemeArrow9pt, kThemeStateActive, NULL, (UInt32) NULL);
 		} /*MyThemeButtonDrawCallback*/
 
 #endif
